docs/embeddings/chefia_embedding/requirements.txt
# ChefIA - Recipe Search Engine Requirements
# Instalar com: pip install -r requirements.txt

# Core Dependencies
google-generativeai>=0.3.0
supabase>=2.0.0
python-dotenv>=1.0.0

# Optional: Web Framework Integration
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0

# Optional: Bot Integration
python-telegram-bot>=20.0
discord.py>=2.3.0

# Optional: Logging and Monitoring
python-json-logger>=2.0.0

# Development Dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Type Hints
typing-extensions>=4.8.0
